Only pull keys from db in lsun for faster cache. (#2544)
* Only pull keys from db in lsun for faster cache.

This pull request inhances the speed of the cache creation for lsun dataset. For the "kitchen_train" the speed was getting slow with cache creation taking more then two hours. This speeds up to cache creation in within minutes. The issue was pulling the large image values each time and dropping them.

For more details on this please refer this issue https://github.com/jnwatson/py-lmdb/issues/195.

* Fixed bug in lsun.py when loading multiple categories

* Make linter happy